+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.21] - 2021-01-05
+### Changed
+- `CodeGenerator`: added ability to specify custom type by format
+
 ## [0.20] - 2021-01-04
 ### Changed
-- `CodeGenerator`: templates: added ability to add marker interface
+- `CodeGenerator`, templates: added ability to add marker interface
 - `pom.xml`: bumped dependency version
 
 ## [0.19.1] - 2020-12-02